Add bounds to mm_malloc (and by extension mm_xmalloc).
On the mm_malloc side, this is simple.  We round up the request to be
representable and then bound it (but not the copy inserted into the
allocation tree).  mm_free already looks up the allocation object
and then doesn't use the bounded address (except in a couple comparisons
I've not bothered to change) and thus requires no modification.